requirements.txt
flasgger==0.9.7.1
Flask==3.0.2
Flask_Cors==4.0.0
openai==1.12.0
Pillow==10.2.0
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.31.1
